Fixed compatibility to the 3.x ns_event command
Fixed NsTclCondObjCmd to be compatible to the 3.x pendant in
way it threats the optional timeout argument. The 3.x reverted
to indefinite (i.e. non-timewait) condvar waits when the
optional "timeout" argument was given as zero.

The 4.x version just exited with NS_TIMEOUT (= 0) in such
cases breaking Tcl scripts written for 3.x.

The corrective measure is to check the passed timeout value
and if == 0, revert to non-timewait condition waits as the
3.x does (did).