(Ns_ModuleLoad): don't use per-thread cache for interp of unknown vserver
* nsd/modload.c (Ns_ModuleLoad): Although the first two arguments
to Ns_ModuleLoad are called server and module, they're really just
two arbitrary strings passed to the module.  The nsdb module for
examples passes the driver name and configuration path.  For this
reason we cannot rely on knowing the current virtual server and
so cannot allocate an interp from the per-thread cache.