Set pipes in the child proxy to close-on-exec
Otherwise we may block forever if we attempt to kill the proxy
running "exec" Tcl command.

Redesigned reaper thread not to block the caller by
waiting on slave proxy processes to exit.

The [ns_proxy handle] command accepts one optional argument
which is interpreted as pool name and only those handles
belonging to the given pool are reported.