Add new command [ns_proxy stop] which kills slaves of the running proxies
Changed format of the result of the [ns_proxy active] to be a
list of lists. Each sublist is in key/value format containing
information about the proxy handle, the pid of the slave, the
start-time and the executing script.

Also, the command now accepts one optional argument: the handle
of the proxy for which to return the status report.

Changed [ns_proxy clear] to require the pool argument. The
[ns_proxy clear] and [ns_proxy stop] commands are now symetric
where the former operates on the idle proxies and the latter on the
running proxies.