Convert Ns_ThreadId() to return a uintptr_t
An uintptr_t is an integral type which is as long as a pointer, so
that it can accurately hold an opaque pthread_t whether it is
implemented as an integer type, a pointer, and on 32/64 bit
machines. This fixes many pointer/integer of different size warnings.

Log messages now print the thread ID in hex, for brevity on
64bit machines.

NB: depending on the architecture of your machine, the ABI may have
changed and require a recompile.