Set cookies with the HttpOnly option by default, making cookie data unavailable to client javascript engines
(on those browsers which support it)

A new -scriptable option is available to ns_setcookie to get the
old behaviour. HttpOnly cookies are not as susceptible to XSS
attacks: secure by default.

The C API has a couple of new flags: NS_COOKIE_SECURE and
NS_COOKIE_SCRIPTABLE, which replace the old 1/0 flag to
Ns_ConnSetCookieEx().