Add Ns_DriverRequestProc callback for non-HTTP drivers; return binary from ns_conn content
The result is the same as for regular filters but this is optional. No need
to hack request anymore, it seemed very weird. All the code that handles
request is now checking for not NULL values before using, mostly url space
routines.

Now, it is valid to have request structure NULL values, nothing will be
called, not handlers will be invoked.

Non HTTP drivers will have to provide requestProc, otherwise nothing will happen and
nobody will notice this driver.