Manage lifetime of tcl synchronization objects without trampling memory
The old scheme with both anonymous and named locks was makinf it difficult
to know when to cleanup. The code was broken -- some memory corruption
and some cases missed.

With this change all locks are named, with anon locks recieving an auto-
generated name. Also, we no longer destroy locks because we can't do it
safely, and there seems little point.