New configure parameter "spread" (for the section server/${server} ) to randomize thread recreation.
Background: Requests are distributed to idle threads via CondSignal,
which delivers subsequent signals to idle threads in a circular manner
(i.e. thread t1, t2, .. tn, then t1 again) determined by the
scheduling policy of the operating system. This round robin behavior
has the consequences, that (a) the idle timeout will never be used on
a sufficiently busy (server since round robin wakes it up regularly),
and (b) the connsperthread value will be the same for every thread
(every thread will get the same numbers of requests).  As a
consequence, multiple threads terminate about at the same time. This
mass termination of threads is performance-wise bad, since these
threads will be restarte as well at about the same time. This is
especially bad when many connections threads are configured
(e.g. 100), since it is likely that about the same amount of threads
will be recreated.

The "spread" parameter introduces a random factor into threadtimeout
and maxconnections to avoid these termination cycles. It adds +/-
the specified percentage to maxconnections and threadtimeout (e.g. a
spread of 20 means a random value between 0.80 to 1.20 of the
specified value of maxconnections and threadtimeout for every
connection thread).