Preallocated (warm) threads, less eager thread creation:
Background: the bookkeeping of available/running threads did not
account for threads already being created. This lead to situations
where on peak loads (e.g. 5 more requests in the queue, all threads
busy an arbitrary number of new threads was scheduled to be created,
causing a heavy burden on the server. Often, when the threads are were
ready and running, the queue situations was already quite different
(often empty). This situation could be as observed as bulky thread
creations and happend e.g. on file-system latencies (e.g. ext3 has
several potential problems in this area).

The changes correct bookkeeping and avoid the overeager thread
creation by sequentialization of thread creates. Furthermore, it
pre-initializes threads such that only initialized threads are in the
waiting list (important, when thread create operations are expensive,
e.g. on large blueprints). This reduces waiting time for clients
esp. on multiprocessor/multicore machines.  The server uses
"minthreads" to control how many pre-initialized connection threads
should be maintained.