New config parameter "concurrentcreatethreshold" to control concurrent thread creation. Background: classical aolserver performs eager thread creation, leading to bursty thread creations and huge resource consumptions (e.g. a short database lock and a constant request stream can cause easily maxthreads to be created without helping to improve the situation. Furthermore, some modules have problems seem to have problem with parallel initializations that can lead to crashes.
Naviserver can now control whether parallel connection thread creations are
wanted or not via the degree (percentage) the request queue is full to allow
cuncurrent connection thread createions.
When concurrentcreatethreshold is to to 0, concurrent thread creation is always allowed;
When concurrentcreatethreshold is to to 100, concurrent thread creation is not allowed;
value inbetween allow parallel thread creations in certain stress situations.