Add option "-all" to ns_getcookie
Background: By using a cookie domain, it is possible that a cookie with a specified name
is given two value, one from e.g. the parent domain, one from the current; this can
lead to situation not easy to debug, when e.g. domain cookies are introduced from
other applications in the same domain. The "-all" option provides instrastructure
support to detect such situations by returning potentially multiple values for the cookie.