nscgi: new config parameter "allowstaticresources"
The new parameter "allowstaticresources" allows to control whether or
not static resources can be delivered from a cgi-bin directory.

Background: when cgi scripts are called from a cgi-bin directory
without an explicit extension mapping to a script interpreter
(e.g. Perl, Bash, ...)  the script has to be set executable. If this
is not the case, the cgi-script is delivered in source code, which
might reveal unwanted information. The reason for this behavior was,
that one can so deliver also static resources (e.g. images) from a
cgi-bin directory. Now this behavior con be controlled with the
configure parameter "allowstaticresources", which is by default off.
If an application depends on this behavior, please turn it on in the
config file (in the nscgi section).