Resolve potential naming conflict
The command NsTclEnvObjCmd() is/was registered twice, once as "ns_env"
and once as "env". Since the latter is not documented, adds no functionality
over "ns_env", does not follow the usual naming conventions and poses
therefore a potential naming conflict, mark it as deprecated.
The C implementation of "env" was replaced by a Tcl one, which will complain.