Added new config parameter for nslog: "logthreadname"
When the parameter is specified, the thread name is placed as second
field in the access.log.  This eases debugging, since one can now link
the lines in error.log caused by an request with the line in the
access.log without the need of guessing. The thread name contains the
the name of the connection pool and a connection id.