return uniform results for "ns_http wait" and "ns_http run" in form of a list of attribute/value pairs (dict)
Background:
Previously, the commands returned different results, depending on the options
- "ns_http run" returned the name of the queued command
  (which was completely useless, this the queued command was already terminated)
- "ns_http wait" returned sometimes the body of the reply (when the reply
  was not spooled) or empty (spooled) or "1", when the option "-R" was specified

Now the result is always a dict, maybe mirroring the result of the (optional)
output variables, which continue to work.

Since the results were irregular and not documented in the man pages,
the backward compatibility should be high. Also the regression test
of NaviSever was not effected.