new API call: ns_db currenthandles
Return information about handles currently allocated to the thread
performing this call. The result contains a dict with the pools,
from which handles were allocated, the names of the handles and their
"active" state (a handle is active between a "ns_db select" and the
last "ns_db getrow" statement). If no handles are allocated by this
thread, empty is returned.

For example, a possible result is "main {nsdb0 1} subquery {nsdb1 0}"
indicating that the current thread has allocated handles from pool
"main" and "subquery", but the handle from pool "main" (i.e. "nsdb0")
is currently active (during a select).