Improve connect() operations with half-configured IPv6 installations
This change adresses especially a problem, where IPv6 is activated on
a server, but it cannot be used to connect to the outside world.
In this situations, connect operations to hosts with domain names
having both IPv4 and IPv6 addresses would fail, although IPv4
connections are possible. In this cases, one should actually
either deactivate IPv6 on the machine or probably deactivate
IPv6 name lookups in bind (option "-4", filter-aaaa-on-v4 yes;)

With this change, this should be be necessary.
This change fixes also a resource leak in this situation.