allow omission of file extension for binary modules,
Background: unixoid platforms use itraditionall in NaviServer
".so" as extension, no matter what "info sharedlibextension" is.
On macOS, the sharedlibextension returns .dylib, while the
module is named .so. Under windows, it seems that
the extension of the module has to be .dll, so it is
not so easy to write portable config files. The new
code tries now first to load the library as specified
(for backward compatibility), if this fails, it
tries to append either ".so" in unixoid platforms
or ".dll" under windows.