Added support for  memory units in config files
With this change, memory sizes con be configured now
not only with bare numbers, but as well with memory
units (e.g. 1MB instead of [expr {1024*1024}]).

The International System of Units (SI) defines
    kB, MB, GB as 1000, 1000^2, 1000^3 bytes,
 and IEC defines
    KiB, MiB and GiB as 1024, 1024^2, 1024^3 bytes.

For effective memory usage, mutiples of 1024 are
better (e.g. 4k pages). Therefore we follow the PostgreSQL
conventions and use 1024 as basis, but we allow as
well the IEC abbreviations.