New feature: Allow a single driver to listen on multiple ip adresses
This feature greatly simplifies setups, where a single server is listening
on multiple IP addresses (e.g. IPv4 and IPv6). Previously, it was
necessary to define separate drivers for these, which need different names
but which are typically configured identically. Furthermore, when the
address is omitted, the server performs a loopup from the nostname
to determine the IP address. Previously, it used the first one,
now it can work with all returned addresses.

To use this feature, simply specify the configure values of "address"
as a Tcl list containing multiple IP addresses,

The change backward compatible, old configure files will continue to work.