New feature: scrypt Password-Based Key Derivation Function (RFC 7914)
The scrypt function is a modern replacement for crypt and bcrypt and
derives secret keys from a secret string.  It is based on memory-hard
functions, which offer added protection against attacks using custom hardware
and GPU arrays. The functions requires the compilation of NaviServer against
OpenSSL 3.0 or newer.

Example from RFC 7914:
   ::ns_crypto::scrypt -secret "password" -salt NaCl -n 1024 -r 8 -p 16

      fdbabe1c9d3472007856e7190d01e9fe7c6ad7cbc8237830e77376634b373162
      2eaf30d92e22a3886ff109279d9830dac727afb94a83ee6d8360cbdfa2cc0640