Changed -timeout semantics for [ns_http run|queue] to cover comm/read/write timeouts. Added -expire option to [ns_http run|queue] to limit the task duration.
The previous -timeout semantics corresponded roughly to the new -expire
option but there were cases where a task might indefinitely exceed the
given timeout. Also... it is far more common to encounter/handle read/write
timeouts and less need to limit the task handling as a whole.

This new -timeout/-expire separation allows for both to be adjusted.

Scripts already using the -timeout will most probably not see any
change unless they are handling huge up/downloads. In such cases
we will now continue with the request, instead of terminating it
in the middle. If somebody explicitly needs to limit the task
duration as a whole, he should consider using -expire option.