Implemented OCSP Stapling for NaviServer.
still todo: configure verbsosity, duration of validity

NaviServer performs two level of caching: in-memory caching and disk caching. When the server receives the first TLS request with OCSP stapling turned on, it checks for an already retieved OCSP response. The disk cache file is saved in the "log" directory of the server and uses the serial number of the certificate to be checked as file name with ".der" as extension. When the disk cache file does not exist, an http/https request is made to the server issuing the servers certificate as defined by the Authority Information Access (AIA) Extension. The names of the file and the http/https request for the OCSP response can be obtained from the system log of the server:.