Improved performance of nsv_lappend signficantly
The old implementation showed a linear slowdown for
nsv_lappend operations depending on the number of
elements in the list. The new implementation
avoids frequent string to list and vice versa
operations by using Tcl_DStrings. It is essentially
independent on the number of elements in the nsv.
For a list with 10K elements, the performance
improvment is about a factor of 400.

Test:
   nsv_set d 1 1
   time {nsv_lappend d 1 1} 10000

OLD:
   403.7255 microseconds per iteration

NEW:
   0.860547 microseconds per iteration