ns_cache: base cache size calculation on the full size of a cache entry.
In previous versions, the cache size calculation was based only on the
value size, leaving the size of the key and the size of the overhead
(size of cache entry, about 126 bytes on a 64bit machine) aside. This
can lead to misleading interpretations, when the cache size is 0 bytes
(just using the cache key as a flag), leaving admins wondering about
the memory consumption.