Added support for "100 CONTINUE"
This feature is in HTTP since a very long time, but not widely supported
although useful (best support is currently in curl). When the client
sends "Expect: 100-continue", the server can check the header fields and
provide an early feedback, whether it is useful to transmit the full
body. This way, e.g. content which is too large can be rejected quickly.