Improve liveliness under high load (improved handling of connection queue overrun)
When NaviServer processes an HTTP request, it is received (head and
body) and associated with a connection thread pool. If the thread pool
has a thread free, the request is directly assigned to this
thread. Otherwise, the request is added to a queue of the connection
pool, waiting for a connection thread.  Since the queue has a limited
size (limited by "maxconnections"), this queue might overrun. In this
situation, the received request is added to a waiting list. Here is,
where the behavior is now improved.

1) previously, as soon as there were entries in the waiting list, no
   more fresh requests were accepted by the driver (just working on
   waiting requests). While this behavior is sensible if only one
   connection pool is available, it causes a blocking behavior in
   cases, when a single connection pool overruns but others could
   continue. Now, more fresh requests are still accepted in this
   situation.

2) In addition, it is now possible to configure the overrun behavior
   via two pool parameters:
   a) pool parameter "rejectoverrun": when set to true, NaviServer sends on
      queue overruns an error "503 service unavailable". The overflowing
      request is not added to the waiting list.
   b) when (a) is set, it is possible to specify parameter "retryafter"
      with a seconds interval to add reply header field "Retry-After" to the
      503 response. (see https://tools.ietf.org/html/rfc7231#section-7.1.3)

When "rejectoverrun" is false (default) then behavior is similar to
before, accept the server does not fully block. The request is kept in
the waiting list, which is checked, whenever a request finishes. For
the client, there is no need to resubmit the request, which will be
proceed eventually. Technically, this means an arbirary large queue.

When "rejectoverrun" is true, then the request causing the overrun is
rejected, and will have to be resubmitted by the client (maybe
automatically, when "Retry-After" is specified. However, the client
support for "Retry-After" is rather bad.

For large sites, "rejectoverrun" is probably the better behavior, since
no background queuing in the waiting list is performed. For now, we
keep the old behavior as default for better backwards compatibility.
The default might change in the future.