make behavior of nsv_dict more orthogonal to the rest
instead of providing the "-varname" option to exists, we provide
it to "nsv_dict get". This makes it logically consistent
with the optional last argument in "nsv_dict getdef" (simliar to "nsv_get"
or "ns_cache_get". We can't use the optional last argument here
because of the var-args semantics of the mirrored tcl commands.