Added option for "oauth1" encoding the ns_urelencode
One case now specify "oauth1" for the encoding table of ns_urlencode
to obtain an encoding as specified in RFC 5849 Section 3.6
for the construction of the signature base string and the "Authorization"
header field.

This is not terribly important but better and faster than serveral "manual"
implementation in case oauth1 is needed.