New option for improved resource management: "ns_ictl maxconcurrentupdates ?max?"
This experimental option allows a server admin to configure, how many
concurrent interpreter updates induced via epoch changes (e.g. on
ns_eval) are allowed. This option is just useful for large sites with
high number of connection threads defined. The default value for this
parameter is sufficiently high, such that all updates will be
performed by default potentially concurrent.

Background: For example, when a sites allows unlimited parallel
updates, and the site has defined e.g. 100 connection threads, and
every single update takes 1 second, then the request processing of the
whole server comes to a complete standstill on epoch increments for at
least for 1 second, since all threads will be busy with updates. When
the number is sufficintly high it all available cores will be used,
which will slow down the concurrent interpreter updates further. When
the server will receive a higher number of requests (e.g. 800 per
second), queueing will be inevitable (unless the number of concurrent
updates is adjusted).

When during an update the maximum number of concurrent updates is
reached, further updates are delayed until the count is again below
this threshould. In such cases it is possible that some request will
be still use the previous blueprint (similar to "eventually
consistent" in distributed database systems).