Reduced time sensitivity on time format when rolling log files
Rolling log files happens often at midnight, using often a day
precision. When e.g. a scheduled procedure the time when this function
is called might be slightly after the scheduled time, which might lead
to a day jump. The problem aggrevates, when multiple log files are
rotated.  The new code identifies day wraps and uses in such cases the
earlier day.