Increase sysadmin-friendliness in error cases.
In case OCSP is activated and the connection to the AIA-server is
blocked, do not reject all traffic but behave like in a configuration
without having OCSP activated, and complain the system log.
Background: Sometimes requests to the let' encrypt AIA server seems to
be blocked by firewalls.

Also, avoid calls to functions recently deprecated in OpenSSL 3.0.