Base repeated scheduled procs on last scheduled time rather than on last finish time
Background: previously, the scheduling of repeated threads was based on
the last finish time. This has the consequence that the execution time
might drift away more and more depending on the runtime of certain
jobs. If one has e.g. a service that should run every minute, some of
these minutes might be skipped by the cumulative drift on a long
running server.