Fix Ns_StrTrimRight() to avoid damaging of UTF-8 characters
Background: Up to now, NaviServer was using "CHARTYPE(space, c)" to
determine, which characters can be trimmed at the end of a string.
Unfortunately, there exists characters, which are classified as
"space", but which can be trailing bytes with different semantics
in multibyte UTF-8 characters (e.g. 0x85). When these bytes are stripped
the result are invalid UTF-8 characters.