check validity of UTF request internally.
This test can only be performed, after the server is determined, since different
servers might have different URLencodings (which is probably an overkill).
Note that Tcl transforms sometimes invalid UTF-8 into valid UTF-8, such
as e.g. /x%c3. Not sure, if this is desired behavior or this should as
well be handled in nsd.