nsproxy: Disambiguate the name of the helper cmd with the name of the module
Background: a bug report stated that
  Error: modload: /usr/local/ns-4.99.21/bin/nsproxy: couldn't load file "/usr/local/ns-4.99.21/bin/nsproxy": /usr/local/ns-4.99.21/bin/nsproxy: cannot dynamically load position-independent executable
on newer Linux versions. It turned out that there are indeed changes with PIE
in glibc(>=2.30). But the real problem was that the nsproxy cmd
was tried to be loaded as a module. To avoid any possible
confusion and to ease debugging, the helper command is
now called "nsproxy-helper" rather than "nsproxy", and
the name of the command can be specified via Makefile.