Factor out WaitForConnect() as own function and extend its behavior
Background: NaviServer performs async connects, which are apparently
difficult to handle, especially when an DNS name is bound to multiple
IP addresses, where some of those cannot be used. The usual
recommendation is to use after async connections in EINPROGRESS cases
getsockopt(fd, SOL_SOCKET, SO_ERROR, ...) to determine
errors. However, this returns *sometimes* false positives. When
someone checks after a connect operation if the socket is writable,
this fails also as a detector for an existing connection, because at
least new Linux versions return the writable flags in EINPROGRESS
cases via poll().  So, an additional method was added based on
getpeername(), which tries to retrieve the peer address for a file
descriptor.