new command: ns_parsehostport ?-strict? string
This command exposes in essence the C-level API function Ns_HttpParseHost2()
to the Tcl level and performs a part of the parsing performed by
"ns_parseurl". In essence, it provides IP-literal parsing, parsing of
the optional port and validation of the characters allowed in the host name.

In addition, the parsing of ns_parseurl became more strict, since
the RFC requires an ALPHA character as first character of the scheme.